fix(transfers)!: make transfer status schemas compatible with TransferStatus enum (#136)
compile-time errors were emitted for transfer responses that used
TransferStatus.TransferComplete for a status (for example)

BREAKING CHANGE: removes TransferInStatusPreTx, TransferInStatusPostTx,
and TransferOutStatus. This is not strictly necessary, but to keep them
we'd need to maintain parallel lists of transfer statuses between them
and the zod schemas, which could lead to bugs. If one of these enums was
needed for typing, they could be replaced with `z.infer<typeof
transferInStatusPostTxSchema>`, for example. If one of the enums was
needed for iterating over, one could use
`transferInStatusPostTxSchema.enum` to get an iterable for the set of
transfer statuses that were in the enum.